rename app theme to palette
