+++ lib/views/screens/authentication/widgets/intro_section.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../utils/app_theme.dart';
+import '../../../../utils/palette.dart';
 
 class IntroSection extends StatelessWidget {
   final String? title;
