+++ lib/views/screens/chats/messages/widgets/file_picked_widget.dart
@@ -1,7 +1,7 @@
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../../utils/app_theme.dart';
+import '../../../../../utils/palette.dart';
 import '../../../../../utils/file_utils.dart';
 
 class FilePickedWidget extends StatelessWidget {
