+++ lib/views/screens/chats/messages/widgets/image_preview_screen.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:image_picker/image_picker.dart';
 
-import '../../../../../utils/app_theme.dart';
+import '../../../../../utils/palette.dart';
 import '../../../../../utils/app_utils.dart';
 import '../../../../../utils/exception.dart';
 import '../../../../../utils/image_utils.dart';
