+++ lib/views/screens/connections/search/widgets/connection_status_widget.dart
@@ -5,7 +5,7 @@ import '../../../../../blocs/bloc.dart';
 import '../../../../../model/app_enum.dart';
 import '../../../../../model/invitation_model.dart';
 import '../../../../../model/pure_user_model.dart';
-import '../../../../../utils/app_theme.dart';
+import '../../../../../utils/palette.dart';
 import '../../../../../utils/navigate.dart';
 import '../../../chats/messages/messages_screen.dart';
