+++ lib/views/screens/connections/tabs/connections/connections_network.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:pure/utils/app_theme.dart';
+import 'package:pure/utils/palette.dart';
 
 import '../../../../../blocs/bloc.dart';
 import '../../../../../services/connection_service.dart';
