+++ lib/views/screens/connections/tabs/connections/invitations/widgets/inviter_profile.dart
@@ -5,7 +5,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../../../../../../blocs/bloc.dart';
 import '../../../../../../../model/inviter_model.dart';
 import '../../../../../../../model/pure_user_model.dart';
-import '../../../../../../../utils/app_theme.dart';
+import '../../../../../../../utils/palette.dart';
 import '../../../../../../../utils/app_utils.dart';
 import '../../../../../../../utils/navigate.dart';
 import '../../../../../../widgets/avatar.dart';
