+++ lib/views/screens/home/home_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../utils/app_theme.dart';
+import '../../../utils/palette.dart';
 import '../../widgets/bottom_bar.dart';
 import '../../widgets/page_transition.dart';
 import 'posts/create_post_screen.dart';
