+++ lib/views/screens/settings/profile/mutual_connection_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../../../blocs/bloc.dart';
 import '../../../../model/invitation_model.dart';
 import '../../../../model/pure_user_model.dart';
-import '../../../../utils/app_theme.dart';
+import '../../../../utils/palette.dart';
 import '../../../../utils/navigate.dart';
 import '../../../widgets/avatar.dart';
 import '../../../widgets/shimmers/loading_shimmer.dart';
