+++ lib/views/widgets/custom_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../utils/app_theme.dart';
+import '../../utils/palette.dart';
 
 class CustomOutlinedButton extends StatelessWidget {
   const CustomOutlinedButton({
