+++ lib/views/widgets/message_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:pure/utils/app_theme.dart';
+import 'package:pure/utils/palette.dart';
 
 class MessageDisplay extends StatelessWidget {
   final String? title;
